added Stop function to stopwatch
When the stopwatch has started, the start button changes to stop, which
stops the stopwatch. The save button works only when the stopwatch is
stopped and there have been at least one recorded lap. When the
stopwatch starts again after being stopped it's timer  and the recorded
laps are reset.